makedoc.py works properly now (really, I mean it this time)]
I have completely re-written makedoc.py. It still uses the inspect module, but has one recursive parse method for everything, rather than separate ones. It's a bit of a pig, because I have to stop it trying to get into the standard library, and getting stuck in recursive loops where it reads to code from classes etc etc.

It does work properly now, and indents nicely. It produces markdown compatible html, because otherwise markdown destroys underscores for italics. It has a nice link to the contents, that permanently loiters in the top right hand corner.

Sorry for all the fuss and broken versions,
Colin Rothwell