Add redis-based session store
With this commit I would like to introduce a new session store based on Redis.  This store creates a sorted set with key `webpy.sessions` containing all the id of all the active sessions, and a key with name`webpy.session:$ID` containing the data associated with the active session.