Python3 migration: db.py tests now run perfectly with Py3
Py2 : tests runned with all DB backends, excepted psycopg1 (which is not
supported anymore and should be removed from web.py as well)
Py3 : tests runned with psycopg2, sqlite3 and mysqlclient (the latter
seems to be compatible with all MySQLdb-related code. MySQLdb is not
available under Py3, and mysqlclient seems at the moment the best
alternative).