Switched to setuptools and bumped verson to 0.40.dev0
Following the legacy of web.py versioning the changes to first digit after the first decimal indicates major change. 0.3x to 0.4x is a major release, possibly backward incompatible.

Now are adding support for Python 3 and dropping support for older versions of Python, so this is a major release and we are bumping the version to 0.40.

It would be nice to switch to semaver, may be for 1.0.