Fix: enable Python 3.12 to parse templates (#785)
* Fix: enable Python 3.12 to parse templates

Closes #784.

This PR fixes the inability of `web.py` to run on Python 3.12, and it
does so by matching any unmatched `"` in a string.

In Python 3.12, changes to `tokenize.generate_tokens()` requires all
quotes to be matched, and the `web.py` template parser hands
unterminated strings to `tokenize.generate_tokens()` quite frequently.

This PR makes a tacit assumption that we should avoid rewriting the
template parser if possible while also allowing `web.py` to run on
Python 3.12.

To that end, it handles every `TokenError` that is because of unmatched
string literals by adding a `"` to the line. Although this appears to
work, I am aware that appending a `"` to each line that has an unmatched
`"` is less than ideal. However, it seemed a very easy way to avoid
rewriting the template parser. I am absolutely open to other approaches.

*IF* this approach looks as if it has legs, I can can look more into
where that "extra" `"` goes, why it seems not to matter, and write some
unit tests.

With regard to testing, I tested this with as many pages on Open Library
that I could (via the local development environment), and it seems to
work.

* noqa: complexity and excess statements for read_expr()

* Add suggestions from @tfmorris.

* Switch to more-itertools.peekable

 and set linter limits back down again

* Fix off-by-1 error and add test coverage

* Linting fixes

* Add `more_itertools` as a dependency

`more_itertools.peekable()` is used now.

`cheroot` already requires `more_itertools`, so this depedency is
already installed anyway, but adding it as a dependency for `webpy`
itself will ensure that if `cheroot` drops the `more_itertools`,
dependency, `webpy` will still require it.

---------

Co-authored-by: Tom Morris <tfmorris@gmail.com>