Fix: disable accessing special variables from template (#803)
* Fix: disable accessing special variables from template

Before this fix, it was possible to import modules from templates, which opens
the possibility of RCE. This applies only when templates from untrusted users
is allowed in the application. While this is not a common scenario, given the
design goals of web.py templating system to be sandboxable, this is a serious
flaw.

This commit fixes that issue by rejecting access to special variables with
names starting with `__` at the compile time.

Fixes #801.